+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 CHANGELOG
 *********
 
+`v2.2.2`_ (2021-05-03)
+======================
+* Fix issue sissaschool/xmlschema#243 (assert with xsi:nil usage)
+* First implementation of XPath 3.0 fn:format-integer
+
 `v2.2.1`_ (2021-03-24)
 ======================
 * Add function signatures at token registration
@@ -305,3 +310,4 @@ CHANGELOG
 .. _v2.1.4: https://github.com/sissaschool/elementpath/compare/v2.1.3...v2.1.4
 .. _v2.2.0: https://github.com/sissaschool/elementpath/compare/v2.1.4...v2.2.0
 .. _v2.2.1: https://github.com/sissaschool/elementpath/compare/v2.2.0...v2.2.1
+.. _v2.2.2: https://github.com/sissaschool/elementpath/compare/v2.2.1...v2.2.2